fix slice sprite when clipped via import max size
