+++ Assets/Editor/Scripts/SpriteSlicer.cs.meta
